Add jar file of this project
maze.jar is the executable file of Laser Maze project. 
To run maze.jar in the command line, switch to the path where maze.jar is in, and type:
maze.jar path_to_inputfile/input_file_name path_to_output_file/output_file_name
If there are white spaces in the path and the filename, use double quote to quote the “file_path/file_name”. Please use forward slash "/" in your path. 
maze.jar will solve the Laser Maze game using information given in input file and functions in the lasermaze package, and write the output into the output file. The requirement of input and output file are listed in the instruction document - DataScientistHW-LaserMaze3_updated1.pdf. If the input file is not in the required format, no output file will be generated.